Initial Pass at adding new Lighted Objects and Color Sets
Moving from color palettes to color sets so that we can define an arbitrary number of unique colors and not get blends between those colors.

Adding lighted objects and a light display to own them.  This will allow us to move to a model where you can create a new lighted object and configure custom effects on that object.