Switch from strip to light display
This is switching the code from using the old wled implementation of strip to the new lightdisplay implementation.  I have needed to introduce a few defines to disable code (see MDR TEMP tags in code) and I have removed some functionality to get things mostly working.  This should be a good starting point for implementing color sets and effects and scenes for the new wled implementation.